pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strands-sglang"
version = "0.2.7"
description = "SGLang model provider for Strands Agents SDK with Token-in/Token-out support for agentic RL training."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    { name = "Yuan He", email = "yuanhe.cs.ai@gmail.com" }
]
keywords = ["strands", "agents", "sglang", "llm", "ai", "reinforcement-learning", "rl", "agentic-rl"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "strands-agents[openai]",
    "aiohttp>=3.8.0",
    "transformers>=4.0.0,<5.0.0",
    "jinja2",
]

[project.urls]
Homepage = "https://github.com/horizon-rl/strands-sglang/"
Documentation = "https://github.com/horizon-rl/strands-sglang/#readme"
Repository = "https://github.com/horizon-rl/strands-sglang/"
Issues = "https://github.com/horizon-rl/strands-sglang/issues"

[project.optional-dependencies]
vision = ["transformers[vision]"]  # also requires torch + torchvision, installed separately
dev = [
    # Testing
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    # Linting & Build
    "pre-commit",
    "ruff>=0.1.0",
    "twine>=4.0.0",
    "build>=0.10.0",
    # Development
    "ipykernel>=7.1.0",
    # Agent-level testing
    "strands-agents-tools",
]

[tool.hatch.build.targets.wheel]
packages = ["src/strands_sglang"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = ["integration: tests that require a running SGLang server"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
